+++ src/main/java/org/purescript/parser/PSParserDefinition.kt
@@ -16,6 +16,7 @@ import org.purescript.file.PSFileStubType
 import org.purescript.lexer.PSLexer
 import org.purescript.psi.*
 import org.purescript.psi.`var`.PSVar
+import org.purescript.psi.import.*
 
 class PSParserDefinition : ParserDefinition, PSTokens {
     override fun createLexer(project: Project): Lexer {
