+++ src/main/java/org/purescript/psi/import/PSImportAlias.kt
@@ -1,8 +1,10 @@
-package org.purescript.psi
+package org.purescript.psi.import
 
 import com.intellij.lang.ASTNode
 import com.intellij.psi.PsiElement
 import com.intellij.psi.PsiNamedElement
+import org.purescript.psi.PSProperName
+import org.purescript.psi.PSPsiElement
 
 /**
  * The alias of an import declaration.
@@ -16,7 +18,8 @@ import com.intellij.psi.PsiNamedElement
  */
 class PSImportAlias(node: ASTNode) : PSPsiElement(node), PsiNamedElement {
 
-    private val properName: PSProperName get() =
+    private val properName: PSProperName
+        get() =
         findNotNullChildByClass(PSProperName::class.java)
 
     override fun setName(name: String): PsiElement? {
