+++ src/main/java/org/purescript/psi/import/PSImportList.kt
@@ -1,8 +1,9 @@
-package org.purescript.psi
+package org.purescript.psi.import
 
 import com.intellij.lang.ASTNode
 import com.intellij.psi.PsiElement
 import org.purescript.parser.PSTokens
+import org.purescript.psi.PSPsiElement
 
 /**
  * The import list in an import declaration.
