+++ src/main/java/org/purescript/psi/import/PSImportedDataMember.kt
@@ -1,5 +1,6 @@
-package org.purescript.psi
+package org.purescript.psi.import
 
 import com.intellij.lang.ASTNode
+import org.purescript.psi.PSPsiElement
 
 class PSImportedDataMember(node: ASTNode) : PSPsiElement(node)
